Day29/requirements.txt
requests
streamlit-tags
streamlit-aggrid
streamlit-option-menu